requirements.txt
accelerate
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
color-extraction==0.1a4
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
datasets
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
dnspython==2.6.1
easy-vqa==1.0
einops==0.8.0
email_validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.1
# flash-attn==2.6.3
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gradio
gradio_client
gradio_imageslider
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub
idna==3.6
imageio==2.34.0
importlib_metadata==8.2.0
importlib_resources==6.3.2
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
openpyxl==3.1.2
orjson==3.10.3
overrides==7.7.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.18.0
pymatreader==0.0.32
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.1.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers
shapely==2.0.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
terminado==0.18.1
threadpoolctl==3.4.0
tifffile==2023.7.10
timm
tinycss2==1.2.1
tokenizers
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
torchdata==0.7.1
torchtext==0.17.1
torchvision==0.17.1
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
git+https://github.com/huggingface/transformers.git
triton==2.2.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
widgetsnbextension==4.0.10
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.20.0
pytorch-lightning==2.4.0
sentencepiece
zstandard
git+https://github.com/LLaVA-VL/LLaVA-NeXT.git
tensorboardX
tensorboard
albumentations
wordcloud
num2words